Updated Order List

+++ FoodOrderApp/app/src/main/java/com/unitarcafe/hegaa/unitarcafe/Database/Database.java
@@ -37,7 +37,7 @@ public class Database extends SQLiteAssetHelper{
                result.add(new Order(
                        c.getString(c.getColumnIndex("ProductName")),
                        c.getString(c.getColumnIndex("Quantity")),
-                       c.getInt(c.getColumnIndex("Price")),
+                       c.getDouble(c.getColumnIndex("Price")),
                        c.getInt(c.getColumnIndex("Discount"))
                        ));
             }while (c.moveToNext());
